requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-mgmt-costmanagement
azure-identity
azure-mgmt-resource>=18.0.0
azure-functions-durable